requirements.txt
Pillow>=10.2.0,<11.0.0 